[fix] #57 리액션 오류 수정

+++ data/src/main/java/com/dpm/sixpack/data/source/remote/dto/response/FinishRunningResponseDto.kt
@@ -31,9 +31,9 @@ data class FinishRunningResponseDto(
 ) {
     fun toRunningSessionResult() =
         RunningSessionResult(
-            totalDistanceMeter = totalDistance,
+            totalDistanceMeter = totalDistance.toInt(),
             totalDurationSec = totalDuration,
-            avgPace = avgPace,
+            avgPace = avgPace.toInt(),
             avgCadence = avgCadence,
             maxCadence = maxCadence,
         )
